Refactoring update to an external class
